% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sightings.R
\docType{data}
\name{clefairy}
\alias{clefairy}
\title{All Clefairy Sightings in Taipei City}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3385 rows and 5 columns.}
\usage{
clefairy
}
\description{
\describe{
 \item{pokemonId}
 \item{longitude}
 \item{latitude}
 \item{created}
 \item{dup_cnt}{duplication count, ignore
 this column since it was used to rule out the data
 duplication during the collecting process.}
}
}
\keyword{datasets}

