% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sightings.R
\docType{data}
\name{daily_pm_count}
\alias{daily_pm_count}
\title{Daily Pokemon Sightings Count}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 990 rows and 3 columns.}
\usage{
daily_pm_count
}
\description{
Pokemon daily sightings in Taiwan between
sampling period.
}
\keyword{datasets}

