% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sightings.R
\docType{data}
\name{pm_count_by_hour}
\alias{pm_count_by_hour}
\title{Hourly Pokemon Sightings Count}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3353 rows and 3 columns.}
\usage{
pm_count_by_hour
}
\description{
Derived from daily_pm_count
}
\details{
\describe{
 \item{pokemonId}
 \item{hour_created}{UTC+8}
 \item{cnt}{count}
 }
}
\keyword{datasets}

