% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sightings.R
\docType{data}
\name{pm_freq}
\alias{pm_freq}
\title{Pokemon Sightings Frequencies}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 142 rows and 3 columns.}
\usage{
pm_freq
}
\description{
Derived from daily_pm_count
}
\keyword{datasets}

