% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sightings.R
\docType{data}
\name{pm_freq_by_hour}
\alias{pm_freq_by_hour}
\title{Hourly Pokemon Sightings Frequencies}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3353 rows and 4 columns.}
\usage{
pm_freq_by_hour
}
\description{
Derived from pm_count_by_hour
}
\details{
\describe{
 \item{pokemonId}
 \item{hour_created}{UTC+8}
 \item{cnt}{count}
 \item{freq}{frequencies}
}
}
\keyword{datasets}

