% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sightings.R
\docType{data}
\name{sightings_grid_latlon}
\alias{sightings_grid_latlon}
\title{Pokemon Sightings in Taiwan Grouped in Grid Cells}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 844 rows and 5 columns.}
\usage{
sightings_grid_latlon
}
\description{
Average daily sightings in Taiwan between
sampling period.
}
\details{
\describe{
 \item{cell_x}
 \item{cell_y}
 \item{avg_cnt}{average count}
 \item{x}{longitude}
 \item{y}{latitude}
 }
}
\keyword{datasets}

