% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sightings.R
\docType{data}
\name{snorlax}
\alias{snorlax}
\title{All Snorlax Sightings in Taiwan}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6550 rows and 5 columns.}
\usage{
snorlax
}
\description{
\describe{
 \item{pokemonId}
 \item{longitude}
 \item{latitude}
 \item{created}
 \item{dup_cnt}{duplication count, ignore
 this column since it was used to rule out the data
 duplication during the collecting process.}
}
}
\keyword{datasets}

